logout moved to nav
